refactors Portfolio/Projects
